rename legacy/trading and legacy/lending package
