+++ legacy/trading/tradingstate/chain_context.go
@@ -7,10 +7,10 @@ import (
 
 // ChainContext provides the subset of blockchain state that the TomoX legacy
 // engine needs during order matching. It is intentionally minimal — only the
-// methods that legacy/tomox actually calls are included.
+// methods that legacy/trading actually calls are included.
 //
-// Both the core-level TradingEngine interface and the legacy/tomox engine
-// use this type to avoid coupling legacy/tomox to the consensus package.
+// Both the core-level TradingEngine interface and the legacy/trading engine
+// use this type to avoid coupling legacy/trading to the consensus package.
 type ChainContext interface {
 	// CurrentHeader retrieves the current head header of the canonical chain.
 	CurrentHeader() *types.Header
