+++ legacy/trading/tradingstate/common.go
