+++ legacy/trading/tradingstate/database.go
