+++ legacy/trading/tradingstate/encoding.go
