+++ legacy/trading/tradingstate/journal.go
