+++ legacy/trading/tradingstate/orderitem.go
