+++ legacy/trading/tradingstate/state_orderItem.go
