+++ legacy/trading/tradingstate/state_orderbook.go
