+++ legacy/trading/tradingstate/trade.go
