+++ legacy/trading/tradingstate/trading_trie.go
